requirements.txt
pydriller
pandas
gitpython
pygount